Frame correction and "unrecorded" for input display
To make things clearer, I use a red line instead of a red rectangle in
the input display. And the word "unrecorded" to say that the input of
the present frame wasn't recorded yet.

The current frame in REC or read only modes should be equal if they are
in the same game state. So, there's no confusion if the player changes
the modes and makes a comparison.