New function to draw text and more
draw_text: doesn't accept string as coordinates anymore, but has a much
better method of drawing, using a reference instead.

The default text doesn't have a full background, but an outline. That
can be changed in the options.

New draw text function is much simpler in some parts of the code.

There's a new lag indicator that flags a lag 1 frame before. So, you can
know if your input will be taken into account without needing to advance
the frame. This is experimental and should work in 99% if the cases. The
old lag flag is kept though.