Create a new movie editor from lsnes.lua
Commit:
https://github.com/rodamaral/amaraticando-tas/commit/6154e034d534ba8a6d7077b69a49ce2d6d6e3eaa

Most of the code was copied and pasted, with few adjustments.

Use the mouse to edit the input (only in readonly mode and not the
past). This currently shows all controllers for both ports. Cannot
display/edit axis and raxis yet, but SMW doesn't use them anyway.
Subframes can be edited.

This is an alfa version! So beware.